requirements.txt
altgraph==0.17.2
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
async-lru==1.0.2
asyncio==3.4.3
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.2
bottle==0.12.18
bottle-websocket==0.2.9
bs4==0.0.1
button==0.0.3.post3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
DateTime==4.3
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.9
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
drawing==0.0.3
EasyProcess==0.3
Eel==0.14.0
entrypoint2==0.2.1
executing==0.8.3
future==0.18.2
geojson==2.5.0
gevent==20.9.0
gevent-websocket==0.10.1
graphviz==0.14.2
greenlet==0.4.17
h11==0.12.0
hupper==1.10.2
idna==2.10
imutils==0.5.4
install==1.3.4
interaction==1.3
ipython==8.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==2.11.2
Js2Py==0.71
llvmlite==0.34.0
lxml==4.6.3
map==1.3.0
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
MouseInfo==0.1.3
mss==6.0.0
numba==0.51.2
numpy==1.19.1
oauthlib==3.2.0
opencv-python==4.5.4.58
outcome==1.1.0
packaging==21.3
parso==0.8.3
PasteDeploy==2.1.1
pefile==2021.9.3
pickleshare==0.7.5
Pillow==7.2.0
pipwin==0.5.1
plaster==1.0
plaster-pastedeploy==0.7
player==0.6.1
prompt-toolkit==3.0.28
psutil==5.8.0
pure-eval==0.2.2
pure-python-adb==0.3.0.dev0
pyaes==1.6.1
PyAudio @ file:///C:/Users/QvaShquai/pipwin/PyAudio-0.2.11-cp38-cp38-win32.whl
PyAutoGUI==0.9.53
pycparser==2.20
pygame==2.0.0.dev10
PyGetWindow==0.0.9
Pygments==2.11.2
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyjsparser==2.7.1
PyJWT==2.3.0
PyMouse==1.0
PyMsgBox==1.0.9
pynput==1.7.4
pyOpenSSL==21.0.0
pyowm==3.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
pyramid==1.10.4
pyramid-jinja2==2.8
PyRect==0.1.4
Pyrogram==1.0.7
pyscreenshot==2.2
PyScreeze==0.1.28
pySmartDL==1.3.4
PySocks==1.7.1
pyTelegramBotAPI==3.7.3
python3-openid==3.2.0
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywebio==1.4.0
pywin32==302
pywin32-ctypes==0.2.0
PyYAML==5.3.1
radius==0.0.4
render==1.0.0
request2==0.2
requests==2.26.0
requests-oauthlib==1.3.1
selenium==4.0.0
six==1.15.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
stack-data==0.2.0
telebot==0.0.3
TgCrypto==1.2.1
tornado==6.1
traitlets==5.1.1
translationstring==1.4
trio==0.19.0
trio-websocket==0.9.2
ttkthemes==3.2.2
twilio==7.3.1
tzdata==2021.5
tzlocal==4.1
ua-parser==0.10.0
undetected-chromedriver==3.1.5.post4
uritemplate==4.1.1
urllib3==1.26.7
user-agents==2.2.0
venusian==3.0.0
wcwidth==0.2.5
WebOb==1.8.6
websockets==10.2
whichcraft==0.6.1
wsproto==1.0.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.interface==5.1.0
